various improvements
includes commits for:
- ERMS-2770: restructure, setup of general entry point
- ERMS-2925: min and max date for datepicker set
