+++ app/grails-app/i18n/messages.properties
@@ -3221,11 +3221,15 @@ myinst.consortiaSubscriptions.provider = Provider
 myinst.consortiaSubscriptions.runningTimes = Running from / to
 
 # ---------- reporting -----------
-myinst.reporting.reset = Reset wizard
+myinst.reporting.reset = Reset LARS
 myinst.reporting.costItems = Cost items
 myinst.reporting.subProp = Subscription property
 myinst.reporting.orgProp = Organisation property
 myinst.reporting.plaProp = Platform property
+myinst.reporting.general = General
+myinst.reporting.members = Members
+myinst.reporting.costs = Costs
+myinst.reporting.subscriptions = Subscriptions
 
 # ---------- members ----------
 members.confirmDelete = Do you really want to delete the selected members?
