+++ app/grails-app/i18n/messages_de.properties
@@ -3050,11 +3050,15 @@ myinst.consortiaSubscriptions.provider = Anbieter
 myinst.consortiaSubscriptions.runningTimes = Laufzeit von / bis
 
 # ---------- reporting -----------
-myinst.reporting.reset = Wizard zurücksetzen
+myinst.reporting.reset = LARS zurücksetzen
 myinst.reporting.costItems = Kosten
 myinst.reporting.subProp = Lizenzmerkmal
 myinst.reporting.orgProp = Einrichtungsmerkmal
 myinst.reporting.plaProp = Plattformmerkmal
+myinst.reporting.general = Allgemein
+myinst.reporting.members = Teilnehmer
+myinst.reporting.costs = Kosten
+myinst.reporting.subscriptions = Lizenzen
 
 # ---------- members ----------
 members.confirmDelete = Wollen Sie die ausgewählten Mitglieder wirklich entfernen?
