+++ app/grails-app/views/reporting/_subscriptionGraphs.gsp
@@ -2,7 +2,7 @@
 <laser:serviceInjection/>
 
 <div class="ui top attached segment" id="chartContainer${entry.id}">
-    <h2>${entry.respondsTo('dropdownNamingConvention') ? entry.dropdownNamingConvention(institution) : entry.name}</h2>
+    <h2>${entry.dropdownNamingConvention(institution)}</h2>
     <a class="closeSegment" data-entry="${genericOIDService.getOID(entry)}">
         <i class="ui icon times"></i>
     </a>
